refactored transmit related methods into module
Moved Agent::Agent methods related to transmitting data into its own
module and file
